More documentation of the id library.
Move the definitions of AttributeKind away from FFI into constants, where they belong.
