+++ rust-src/id/Cargo.toml
@@ -25,6 +25,7 @@ either = "1.6"
 pairing = "0.15"
 num = "0.3"
 derive_more = "0.99"
+thiserror = "1.0"
 
 # only for printing account addresses
 base58check = "0.1"
