+++ rust-src/id/src/test.rs
@@ -2,15 +2,14 @@ use crate::{
     account_holder::*,
     anonymity_revoker::*,
     chain::*,
-    constants::{ArCurve, BaseField, IpPairing},
-    ffi::*,
+    constants::{ArCurve, BaseField, IpPairing, *},
     identity_provider::*,
     secret_sharing::Threshold,
     types::*,
 };
 use crypto_common::{
-    types::{KeyIndex, TransactionTime},
-    KeyPair, *,
+    types::{KeyIndex, KeyPair, TransactionTime},
+    *,
 };
 use curve_arithmetic::Curve;
 use dodis_yampolskiy_prf::secret as prf;
